Added "restored" category!
Literal is going to like this commit.

Phew! Pretty much everything from "Pochette" has been completely restored. Later on I'll verify the capitalization (since PTCUtilities seems to capitalize anything pasted into it), but other than that, everything's good to go!

The only things that couldn't be restored (at least not at the moment) are the original filenames of the "Pochette" programs (they were never given), so I just used the first 8 characters of the filenames of the HTML files for their pages (after "pochette-", of course).